Move non-JS registration search test to its own spec file
